Create CreativeCashDrawSolutionsRegister
I did this code test using a single page of html with JavaScript and JQuery to run the code.  This allows it to fit into a larger solution but also gives the possibility to just output the data to a file instead of displaying on the page like it is now.

For such a small project I thought that the single page solution would work best.  I'm an experienced web developer so I gravitate toward a browser based solution.  Only client side code was really needed in this case, though I have more experience using C# .Net I decided against it for this demo.

For this demo I did get to learn a bit more JavaScript than I had used in the past.  For example I had the idea I wanted to use a dictionary for the currency denominations but had to look to see if JavaScript actually had that capability.  Also, one other thing I learned from this that I will use going forward is that Sublime Text makes a nice IDE for HTML and JavaScript based solutions.