Eric Fink - Completed Cash Register
Initial and complete cut of Eric Fink's cash register for Creative Cash
Draw Solutions.  This feature will calculate the minimal amount of
change to be given to a customer after a transaction has been completed.
If the amount owed by the customer is divisible by 3, coins are chosen
at random.

Please refer to ASSUMPTIONS_NOTES for further notes and any and all
assumptions worked under.